project update

+++ InhouseMarketplaceApplication/src/main/java/com/cg/ima/service/CategoryService.java
@@ -4,9 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.cg.ima.entities.Category;
+import com.cg.ima.exception.CategoryExistsException;
+import com.cg.ima.exception.CategoryNotFoundException;
 import com.cg.ima.repository.ICategoryRepository;
 
 @Service
@@ -18,10 +21,19 @@ public class CategoryService implements ICategoryService {
 	
 	
 	@Override
-	public Category addCategory(Category category) {
+	public Category addCategory(Category category) throws CategoryExistsException {
 		
-		Category cat = categoryRepository.save(category);
-		return cat;
+		Category searchedCategory = categoryRepository.findBySpecificCategoryName(category.getCatName());
+
+		if (searchedCategory != null && category.getCatName().equals(searchedCategory.getCatName())) 
+		{
+			throw new CategoryExistsException("Category with name '"+category.getCatName()+"' already exists!");
+		}
+		else
+		{
+			Category cat = categoryRepository.save(category);
+			return cat;
+		}
 
 	}
 
@@ -34,14 +46,24 @@ public class CategoryService implements ICategoryService {
 	}
 
 	@Override
-	public void removeCategory(int catId) {
+	public ResponseEntity<String> removeCategory(int catId) throws CategoryNotFoundException {
 		
-		categoryRepository.deleteById(catId);
+		Optional<Category> searchedCategory = categoryRepository.findById(catId);
+
+		if (!searchedCategory.isPresent()) 
+		{
+			throw new CategoryNotFoundException("Category ID '"+ catId +"' Doesn't Exists!");
+		}
+		else
+		{
+			categoryRepository.deleteById(catId);
+			return ResponseEntity.ok("Category Id '"+catId+"' has been deleted!");
+		}
 		
 	}
 
 	@Override
-	public Category getCategoryById(int catId) {
+	public Category getCategoryById(int catId) throws CategoryNotFoundException {
 		
 		Optional<Category> cat = categoryRepository.findById(catId);
 
@@ -51,7 +73,7 @@ public class CategoryService implements ICategoryService {
 		} 
 		else 
 		{
-			return null;
+			throw new CategoryNotFoundException("Category ID '"+ catId +"' Not Found!");
 		}
 		
 	}
@@ -67,23 +89,34 @@ public class CategoryService implements ICategoryService {
 
 // if we want a single category by their full name		
 	@Override
-	public Category getCategoryByName(String catName) {
+	public Category getCategoryByName(String catName) throws CategoryNotFoundException {
 		
 		Optional<Category> cat = categoryRepository.findByName(catName);
 
-		if (cat.isPresent()) {
+		if (cat.isPresent()) 
+		{
 			return cat.get();
 		}
-
-		return null;
+		else
+		{
+			throw new CategoryNotFoundException("Category Name '"+ catName +"' Not Found!");
+		}
 		
 	}
 
 
 	@Override
-	public List<Category> getAllCategories() {
+	public List<Category> getAllCategories() throws CategoryNotFoundException {
 		
-		return categoryRepository.findAll();
+		List<Category> categories = categoryRepository.findAll();
+		if (categories.isEmpty()) 
+		{
+			throw new CategoryNotFoundException("No Category Found!");
+		}
+		else
+		{	
+			return categories;
+		}
 		
 	}
