+++ InhouseMarketplaceApplication/src/main/java/com/cg/ima/service/EmployeeService.java
@@ -4,8 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
 import com.cg.ima.entities.Employee;
+import com.cg.ima.exception.EmployeeExistsException;
+import com.cg.ima.exception.EmployeeNotFoundException;
 import com.cg.ima.repository.IEmployeeRepository;
 
 @Service
@@ -17,10 +21,19 @@ public class EmployeeService implements IEmployeeService {
 	
 	
 	@Override
-	public Employee addEmployee(Employee emp) {
+	public Employee addEmployee(Employee emp) throws EmployeeExistsException {
 		
-		Employee employee = employeeRepository.save(emp);
-		return employee;
+		Employee searchedEmployee = employeeRepository.findByEmployeeEmail(emp.getEmail());
+
+		if (searchedEmployee != null && emp.getEmail().equals(searchedEmployee.getEmail())) 
+		{
+			throw new EmployeeExistsException("Employee with email '"+emp.getEmail()+"' already exists!");
+		}
+		else
+		{
+			Employee employee = employeeRepository.save(emp);
+			return employee;
+		}
 
 	}
 
@@ -33,14 +46,24 @@ public class EmployeeService implements IEmployeeService {
 	}
 
 	@Override
-	public void removeEmployee(int empId) {
+	public ResponseEntity<String> removeEmployee(int empId) throws EmployeeNotFoundException {
 		
-		employeeRepository.deleteById(empId);
+		Optional<Employee> searchedEmployee = employeeRepository.findById(empId);
+
+		if (!searchedEmployee.isPresent()) 
+		{
+			throw new EmployeeNotFoundException("Employee ID '"+ empId +"' Doesn't Exists!");
+		}
+		else
+		{
+			employeeRepository.deleteById(empId);
+			return ResponseEntity.ok("Employee Id '"+empId+"' has been deleted!");
+		}
 		
 	}
 
 	@Override
-	public Employee getEmployeeById(int empId) {
+	public Employee getEmployeeById(int empId) throws EmployeeNotFoundException {
 		
 		Optional<Employee> emp = employeeRepository.findById(empId);
 
@@ -50,7 +73,7 @@ public class EmployeeService implements IEmployeeService {
 		} 
 		else 
 		{
-			return null;
+			throw new EmployeeNotFoundException("Employee ID '"+ empId +"' Not Found!");
 		}
 		
 	}
@@ -66,23 +89,33 @@ public class EmployeeService implements IEmployeeService {
 
 // if we want a single employee by their full name		
 	@Override
-	public Employee getEmployeeByName(String empName) {
+	public Employee getEmployeeByName(String empName) throws EmployeeNotFoundException {
 		
 		Optional<Employee> emp = employeeRepository.findByName(empName);
 
-		if (emp.isPresent()) {
+		if (emp.isPresent()) 
+		{
 			return emp.get();
 		}
-
-		return null;
-		
+		else
+		{
+			throw new EmployeeNotFoundException("Employee Name '"+ empName +"' Not Found!");
+		}
 	}
 
 
 	@Override
-	public List<Employee> getAllEmployees() {
+	public List<Employee> getAllEmployees() throws EmployeeNotFoundException {
 		
-		return employeeRepository.findAll();
+		List<Employee> employees = employeeRepository.findAll();
+		if (employees.isEmpty()) 
+		{	
+			throw new EmployeeNotFoundException("No Employee Found!");
+		}
+		else
+		{	
+			return employees;
+		}
 		
 	}
