+++ InhouseMarketplaceApplication/src/main/java/com/cg/ima/service/ResourceService.java
@@ -4,9 +4,11 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.cg.ima.entities.Resource;
+import com.cg.ima.exception.ResourceNotFoundException;
 import com.cg.ima.repository.IResourceRepository;
 
 @Service
@@ -34,14 +36,24 @@ public class ResourceService implements IResourceService {
 	}
 
 	@Override
-	public void removeResource(int resId) {
+	public ResponseEntity<String> removeResource(int resId) throws ResourceNotFoundException {
 		
-		resourceRepository.deleteById(resId);
+		Optional<Resource> searchedResource = resourceRepository.findById(resId);
+
+		if (!searchedResource.isPresent()) 
+		{
+			throw new ResourceNotFoundException("Resource ID '"+ resId +"' Doesn't Exists!");
+		}
+		else
+		{
+			resourceRepository.deleteById(resId);
+			return ResponseEntity.ok("Resource Id '"+resId+"' has been deleted!");
+		}
 		
 	}
 
 	@Override
-	public Resource getResourceById(int resId) {
+	public Resource getResourceById(int resId) throws ResourceNotFoundException {
 		
 		Optional<Resource> resource = resourceRepository.findById(resId);
 
@@ -51,27 +63,41 @@ public class ResourceService implements IResourceService {
 		} 
 		else 
 		{
-			return null;
+			throw new ResourceNotFoundException("Resource ID '"+ resId +"' Not Found!");
 		}
 		
 	}
 
 //if we want all Resources using like query	
 	@Override
-	public List<Resource> getAllResourcesByCategory(int catId) {
+	public List<Resource> getAllResourcesByCategory(int catId) throws ResourceNotFoundException {
 		
-		List<Resource> resource = resourceRepository.findByCategory(catId);
-		return resource;
+		List<Resource> resources = resourceRepository.findByCategory(catId);
+		if (resources.isEmpty()) 
+		{
+			throw new ResourceNotFoundException("Resource For Category Id '"+ catId +"' Not Found!");
+		}
+		else
+		{	
+			return resources;
+		}
 		
 	}
 
 
 
 	@Override
-	public List<Resource> getAllResources() {
+	public List<Resource> getAllResources() throws ResourceNotFoundException {
 		
-		return resourceRepository.findAll();
-		
+		List<Resource> resources = resourceRepository.findAll();
+		if (resources.isEmpty()) 
+		{
+			throw new ResourceNotFoundException("No Resource Found!");
+		}
+		else
+		{	
+			return resources;
+		}
 	}
